+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/search/CertifiedProductBasicSearchResult.java
@@ -3,11 +3,18 @@ package gov.healthit.chpl.domain.search;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonView;
+
 public class CertifiedProductBasicSearchResult extends CertifiedProductSearchResult {
 	private static final long serialVersionUID = -2547390525592841038L;
 	
+	@JsonView({SearchViews.Default.class})
 	private List<String> previousDevelopers; //list of previous developer names that owned the product
+	
+	@JsonView({SearchViews.Default.class})
 	private List<String> criteriaMet; //list of criteria numbers
+	
+	@JsonView({SearchViews.Default.class})
 	private List<String> cqmsMet; //list of cmqs that were met (any version)
 	
 	public CertifiedProductBasicSearchResult() {